Kill the mksmlpre script and introduce globals.def
Add a couple of AWK scripts to generate the data from globals.def rather than
the hackery that was the mksmlpre script. Simplify and kill globals.h since
it is easier to maintain it by the generated global_tbl.h.

While working on the game, rewrite generation of Predef.sml as well.