Enable string.sml in a modified form.
I am not sure what is the right thing to do here, so I added a TODO
about asking John Reppy about how the parsing is supposed to work. The
goal is the obtain the idea behing the parsing and how it ought to work.

With that information in hand, we can then proceed to implement the
code correctly. According to the spec, we still soundly fail the
string.fromSting.sml test case, so it is not as if it is working yet.

I note, however, that SML/NJ also fails here on both accounts. So
we have a case where mosml, smlnj and mlton all disagree on the
implementation.